tax_ai_calculator/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
openai>=1.0.0
python-dotenv>=1.0.0
PyPDF2>=3.0.1
pytesseract>=0.3.10
Pillow>=10.0.0
requests>=2.31.0
